first efforts to get cap'n proto logs going

+++ libMagAOX/app/stateCodes.hpp
@@ -9,6 +9,8 @@
 #ifndef app_stateCodes_hpp
 #define app_stateCodes_hpp
 
+#include <cstdint>
+#include <string>
 
 namespace MagAOX 
 {
@@ -25,7 +27,7 @@ struct stateCodes
    /// The type of the state code.
    /**
      */
-   typedef int stateCodeT;
+   typedef int16_t stateCodeT;
 
 
    /// The numeric codes descrbing an application's state
